all StudOn 5.0 changes
# Solved Conflicts:
#	.gitignore
#	Modules/Course/classes/class.ilCourseMembershipMailNotification.php
#	Modules/Course/classes/class.ilCourseRegistrationGUI.php
#	Modules/Course/classes/class.ilObjCourse.php
#	Modules/Course/classes/class.ilObjCourseAccess.php
#	Modules/Course/classes/class.ilObjCourseGUI.php
#	Modules/Course/classes/class.ilObjCourseListGUI.php
#	Modules/Course/templates/default/tpl.member_export.html
#	Modules/Exercise/classes/class.ilExAssignment.php
#	Modules/Exercise/classes/class.ilExAssignmentGUI.php
#	Modules/Exercise/classes/class.ilExGradesTableGUI.php
#	Modules/Exercise/classes/class.ilExParticipantTableGUI.php
#	Modules/Exercise/classes/class.ilExerciseDataSet.php
#	Modules/Exercise/classes/class.ilExerciseMemberTableGUI.php
#	Modules/Exercise/classes/class.ilExerciseMembers.php
#	Modules/Exercise/classes/class.ilObjExercise.php
#	Modules/Exercise/classes/class.ilObjExerciseGUI.php
#	Modules/Forum/classes/class.ilForum.php
#	Modules/Group/classes/class.ilGroupRegistrationGUI.php
#	Modules/Group/classes/class.ilObjGroup.php
#	Modules/Group/classes/class.ilObjGroupAccess.php
#	Modules/Group/classes/class.ilObjGroupGUI.php
#	Modules/HTMLLearningModule/classes/class.ilObjFileBasedLMGUI.php
#	Modules/LearningModule/classes/class.ilObjContentObject.php
#	Modules/Scorm2004/scripts/buildrte/rte.js
#	Modules/Scorm2004/scripts/questions/question_handling.js
#	Modules/ScormAicc/scripts/SCORM1_2standard.js
#	Modules/Session/classes/class.ilObjSession.php
#	Modules/Session/classes/class.ilSessionOverviewGUI.php
#	Modules/Survey/classes/class.ilObjSurvey.php
#	Modules/Survey/classes/class.ilObjSurveyGUI.php
#	Modules/SurveyQuestionPool/classes/class.SurveyMatrixQuestionGUI.php
#	Modules/Test/classes/class.ilObjTest.php
#	Modules/Test/classes/class.ilObjTestGUI.php
#	Modules/Test/classes/class.ilObjTestSettingsScoringResultsGUI.php
#	Modules/Test/classes/class.ilTestArchiver.php
#	Modules/Test/classes/class.ilTestEvaluationGUI.php
#	Modules/Test/classes/class.ilTestExportGUI.php
#	Modules/Test/classes/class.ilTestGradingMessageBuilder.php
#	Modules/Test/classes/class.ilTestOutputGUI.php
#	Modules/Test/classes/class.ilTestPDFGenerator.php
#	Modules/Test/classes/class.ilTestPassesSelector.php
#	Modules/Test/classes/class.ilTestResultsImportParser.php
#	Modules/Test/classes/class.ilTestScoringByQuestionsGUI.php
#	Modules/Test/classes/class.ilTestScoringGUI.php
#	Modules/Test/classes/tables/class.ilTestParticipantsTableGUI.php
#	Modules/Test/templates/default/test_print.css
#	Modules/Test/templates/default/tpl.il_as_tst_print_test_confirm.html
#	Modules/TestQuestionPool/classes/class.assKprimChoiceGUI.php
#	Modules/TestQuestionPool/classes/class.assOrderingHorizontal.php
#	Modules/TestQuestionPool/classes/class.assOrderingQuestion.php
#	Modules/TestQuestionPool/classes/class.assOrderingQuestionGUI.php
#	Modules/TestQuestionPool/classes/class.assQuestion.php
#	Modules/TestQuestionPool/classes/class.assSingleChoice.php
#	Modules/TestQuestionPool/classes/class.ilObjQuestionPoolGUI.php
#	Modules/TestQuestionPool/classes/class.ilTestQuestionPoolImporter.php
#	Modules/TestQuestionPool/templates/default/tpl.il_as_qpl_question_printview.html
#	Services/Authentication/classes/class.ilAuthBase.php
#	Services/Authentication/classes/class.ilAuthUtils.php
#	Services/COPage/classes/class.ilPageObjectGUI.php
#	Services/COPage/xsl/page.xsl
#	Services/Component/classes/class.ilCachedComponentData.php
#	Services/Container/classes/class.ilContainer.php
#	Services/Container/classes/class.ilContainerSortingSettings.php
#	Services/Database/classes/class.ilDBInnoDB.php
#	Services/Database/classes/class.ilDBMySQL.php
#	Services/GlobalCache/classes/class.ilGlobalCache.php
#	Services/Init/classes/class.ilInitialisation.php
#	Services/Init/templates/default/tpl.view_terms_of_service.html
#	Services/Locator/classes/class.ilLocatorGUI.php
#	Services/Mail/classes/class.ilMail.php
#	Services/Mail/classes/class.ilMailFormCall.php
#	Services/Mail/classes/class.ilMailGUI.php
#	Services/MainMenu/classes/class.ilMainMenuGUI.php
#	Services/Membership/classes/Export/class.ilMemberExport.php
#	Services/Membership/classes/Export/class.ilMemberExportGUI.php
#	Services/Membership/classes/class.ilSubscriberTableGUI.php
#	Services/Object/classes/class.ilObjectCopyGUI.php
#	Services/Object/classes/class.ilObjectCopySelectionTableGUI.php
#	Services/PDFGeneration/classes/tcpdf/config/tcpdf_config.php
#	Services/Registration/classes/class.ilAccountRegistrationGUI.php
#	Services/Registration/classes/class.ilRegistrationSettingsGUI.php
#	Services/Registration/templates/default/tpl.usr_registered.html
#	Services/Repository/classes/class.ilRepUtil.php
#	Services/Taxonomy/classes/class.ilTaxonomyDataSet.php
#	Services/TermsOfService/classes/class.ilTermsOfServiceFileSystemDocument.php
#	Services/Tracking/classes/repository_statistics/class.ilTrObjectUsersPropsTableGUI.php
#	Services/Tree/classes/class.ilMaterializedPathTree.php
#	Services/UICore/classes/class.ilTemplate.php
#	Services/UICore/templates/default/tpl.footer.html
#	Services/User/classes/class.ilPublicUserProfileGUI.php
#	Services/User/classes/class.ilUserDataSet.php
#	Services/WebAccessChecker/classes/class.ilWebAccessChecker.php
#	error.php
#	include/inc.ilias_version.php
#	setup/classes/class.ilClientList.php
#	setup/sql/ilDBTemplate.php
#	templates/default/delos.css
#	templates/default/delos.less
#	templates/default/less/Services/Block/delos.less
#	templates/default/less/Services/Calendar/delos.less
#	templates/default/less/Services/Locator/delos.less
#	templates/default/less/Services/MainMenu/delos.less
#	templates/default/less/Services/PersonalDesktop/delos.less

+++ Services/Utilities/classes/class.ilBenchmark.php
@@ -92,16 +92,31 @@ return;
 		{
 			if (is_array($this->db_bench) && is_object($ilDB))
 			{
+// fau: extendBenchmark - store also the total time to compare it with query time
+				global $ilias;
+				$t1 = explode(" ", $GLOBALS['ilGlobalStartTime']);
+				$t2 = explode(" ", microtime());
+				$diff = $t2[0] - $t1[0] + $t2[1] - $t1[1];
+
+				$ilias->setSetting("db_bench_total_time", $diff);
+// fau.
+
 				$this->db_bench_stop_rec = true;
 
-				$ilDB->manipulate("DELETE FROM benchmark");
-				foreach ($this->db_bench as $b)
+// fau: extendBenchmark - check instant setting, add backtrace
+				if (empty($this->db_enabled_instant))
 				{
-					$ilDB->insert("benchmark", array(
-						"duration" => array("float", $this->microtimeDiff($b["start"], $b["stop"])),
-						"sql_stmt" => array("clob", $b["sql"])
-					));
+					$ilDB->manipulate("DELETE FROM benchmark");
+					foreach ($this->db_bench as $b)
+					{
+						$ilDB->insert("benchmark", array(
+							"duration" => array("float", $this->microtimeDiff($b["start"], $b["stop"])),
+							"sql_stmt" => array("clob", $b["sql"]),
+							"backtrace" => array("clob", $b["backtrace"])
+						));
+					}
 				}
+// fau.
 			}
 			$this->enableDbBench(false);
 		}
@@ -294,16 +309,20 @@ return;
 
 		$this->db_enabled = $ilSetting->get("enable_db_bench");
 		$this->db_enabled_user = $ilSetting->get("db_bench_user");
+// fau: extendBenchmark - get setting for instant saving
+		$this->db_enabled_instant =  $ilSetting->get("enable_db_bench_instant");
+// fau.
 		return $this->db_enabled;
 	}
 
+// fau: extendBenchmark - add parameter to enable instant saving, delete old data
 	/**
 	 * Enable DB benchmarking
 	 *
 	 * @param	boolean		enable db benchmarking
 	 * @param	string		user account name that should be benchmarked
 	 */
-	function enableDbBench($a_enable, $a_user = 0)
+	function enableDbBench($a_enable, $a_user = 0, $instant = 0)
 	{
 		global $ilias;
 
@@ -314,6 +333,8 @@ return;
 			{
 				$ilias->setSetting("db_bench_user", $a_user);
 			}
+			$ilias->setSetting("enable_db_bench_instant", $instant);
+			$this->clearData();
 		}
 		else
 		{
@@ -324,6 +345,7 @@ return;
 			}
 		}
 	}
+// fau.
 
 	/**
 	 * start measurement
@@ -359,9 +381,33 @@ return;
 			$this->db_enabled_user == $ilUser->getLogin() &&
 			!$this->db_bench_stop_rec)
 		{
+// fau: extendBenchmark - add backtrace to benchmark, save instantly (to allow benchmarking of fatal errors)
+			$i = 0;
+			foreach (debug_backtrace() as $step)
+			{
+	            if ($i > 0)
+				{
+					$backtrace .= '['.$i.'] '.$step['file'].' '.$step['line'].': '.$step['function']."()\n";
+				}
+				$i++;
+			}
+			$backtrace .= '['.$i.'] '.$_SERVER['REQUEST_URI'];
+
+			if ($this->db_enabled_instant)
+			{
+				global $ilDB;
+				$ilDB->insert("benchmark", array(
+					"duration" => array("float", $this->microtimeDiff($this->start,  microtime())),
+					"sql_stmt" => array("clob", $this->sql),
+					"backtrace" => array("clob", $backtrace)
+				));
+			}
+
 			$this->db_bench[] =
 				array("start" => $this->start, "stop" => microtime(),
-					"sql" => $this->sql);
+					"sql" => $this->sql,
+					"backtrace" => $backtrace);
+// fau.
 		}
 	}
 
@@ -379,12 +425,35 @@ return;
 		$b = array();
 		while ($rec = $ilDB->fetchAssoc($set))
 		{
+// fau: extendBenchmark - get benchmarks with backtrace
 			$b[] = array("sql" => $rec["sql_stmt"],
-				"time" => $rec["duration"]);
+				"time" => $rec["duration"],
+				"backtrace" => $rec["backtrace"]);
+// fau.
 		}
 		return $b;
 	}
 
+// fau: extendBenchmark - get total and query time
+
+	function getDbBenchTotalTime()
+	{
+		global $ilias;
+		return $ilias->getSetting('db_bench_total_time');
+	}
+
+	function getDbBenchQueryTime()
+	{
+		global $ilDB;
+
+		$set = $ilDB->query("SELECT SUM(duration) dsum FROM benchmark");
+		if ($rec = $ilDB->fetchAssoc($set))
+		{
+			return $rec['dsum'];
+		}
+		return 0;
+	}
+// fau.
 }
 
 ?>
